==> kaggle/birdclef_2021/my_code/local/03-first_models.ipynb <==
from pathlib import Path
from tqdm.notebook import tqdm
import librosa
import librosa.display
import numpy as np
import soundfile
import pandas as pd
import joblib
import tensorflow as tf
import tensorflow.keras as keras


import random
import os

## Parameters
Let's introduce a few convenience variables.

SR = 32_000
DURATION = 5
SEED = 42
EPSILON = 10**(-6)

def seed_everything(seed=SEED):
    random.seed(seed)
    os.environ['PYTHONHASHSEED'] = str(seed)
    np.random.seed(seed)
    #torch.manual_seed(seed)
    #torch.cuda.manual_seed(seed)
    #torch.backends.cudnn.deterministic = True
    tf.random.set_seed(seed)

seed_everything()

PATH_DATASET = Path.home() / "datasets/kaggle/birdclef-2021"
PATH_DATASET.exists()

L_birds = [path.name for path
           in (PATH_DATASET / "train_short_audio").iterdir()]
L_birds[:10]

sorted(L_birds) == L_birds

# L_birds will sometimes sorted, sometimes not, so assign just in case.
L_birds = sorted(L_birds) 

- In `01.ipynb`, `L_birds` does not need to be sorted, well, at least it seems.
- Here, in `03.ipynb`, it needs to be sorted.

**Interesting fact?**: Maybe it's due to **seeding**?

D_label_index = {label: i for i, label in enumerate(L_birds)}
D_label_index

D_index_label = {v: k for k, v in D_label_index.items()}
D_index_label

df_train_soundscape = pd.read_csv("./rich_train_soundscape_labels.csv")
df_train_soundscape.head()

L_useful_features = [
    #"longitude",
    "latitude",
    "month",
    "day",
    "is_test",
]
df_useful_features = df_train_soundscape.loc[:, L_useful_features]
df_useful_features

## Dataset (Features)

is_train = df_train_soundscape["is_test"] == False

(is_train == ~df_train_soundscape["is_test"]).all()

(False == ~df_train_soundscape["is_test"]).all()

is_train = ~df_train_soundscape["is_test"]

df_useful_features["month"].unique()

df_useful_features["day"].unique()

df_useful_features["month_normalized"] = (df_useful_features["month"] - 6.5) / 12
df_useful_features["day_normalized"] = (df_useful_features["day"] - 16) / 31
df_useful_features["latitude_normalized"] = df_useful_features["latitude"] / 90

df_useful_features.loc[:,
    ["month", "month_normalized",
     "day", "day_normalized",
     "latitude", "latitude_normalized"]]

XX_train_normalized = df_useful_features.loc[is_train, ["latitude_normalized", "month_normalized", "day_normalized"]]
XX_test_normalized = df_useful_features.loc[~is_train, ["latitude_normalized", "month_normalized", "day_normalized"]]
XX_train_normalized.shape, XX_test_normalized.shape

XX_train_normalized

XX_test_normalized

XX_train_normalized = XX_train_normalized.values

## Dataset (Images)
**(?)** How large can the training dataset be if we load it entirely into RAM?<br>
**(R)** Let's say

- the `dtype` will be `float32`
- each mels will be of shape `(128, 201)`
- there will be `1920` mels

df_train_soundscape.loc[is_train].shape

random_npy = np.load(df_train_soundscape.loc[0, "npy_path"])
random_npy.dtype, random_npy.shape

4 * np.product(random_npy.shape) * df_train_soundscape.loc[is_train].shape[0]

n_bytes = _
print(f"That's {n_bytes:,} bytes.")

n_MB = n_bytes // (2**20)
print(f"That's {n_MB:,} MB.")

This is relatively small memory consumption. Even when we make them into 3-channel images, that is only less than `600 MB`. I think **we can afford caching it in RAM**.

**(?)** In `kkiller`'s code, the `.npy` are saved in `dtype=uint8` and only before going into model is the `dtype` converted again to `float32`.
Why is that?<br>
**(R)** I guess it is because `kkiller` wants to minimize the storage and reload of `.npy` files.

n_train_instances = df_train_soundscape.loc[is_train].shape[0]
X_train = np.empty((n_train_instances, *random_npy.shape, 3), dtype=np.float32)
X_train.shape

for i, row in enumerate(df_train_soundscape.loc[is_train]):
    print(i)
    print(row)

[s for s in dir(df_train_soundscape) if s.startswith("iter")]

help(df_train_soundscape.itertuples)

# index from 0? No!
df_train_soundscape[is_train].head()

for i, row in enumerate((df_train_soundscape.loc[is_train]).itertuples()):
    if i < 3:
        print(i)
        print(row)
    else:
        break

XX_train_normalized.head(3)

help(np.repeat)

**(?)** If `M` is a `(2,2)` ndarray, what do we do if we want three copies of `M` to make up 3 channels like an image?

M = np.array([[1,2], [3,4]])
#np.repeat(M, [2,2])
np.stack([M, M, M])

_.shape

np.stack([M]*3)

Wrong shape: we want **channel-last**.

M[np.newaxis]

np.repeat(M[np.newaxis], 3)

np.repeat(M[np.newaxis], 3, axis=0)

_.shape

MM = np.repeat(M[..., np.newaxis], 3, axis=-1)
MM.shape

MM[..., 0]

M[np.newaxis]

np.repeat(M[np.newaxis], 3, axis=-1)

_.shape

np.repeat(M, 3).reshape((*M.shape, 3))

_[..., 0], _[..., -1]

M.reshape(*(M.shape), 3)

ValueError: cannot reshape array of size 4 into shape (2,2,3)

M[..., np.newaxis].reshape(*(M.shape), 3)

ValueError: cannot reshape array of size 4 into shape (2,2,3)

The above: `reshape` is not broadcast.

# fill in X_train
for i, row in enumerate((df_train_soundscape.loc[is_train]).itertuples()):
    mels_i = np.load(row.npy_path).astype(np.float32, copy=False)
    X_train[i] = np.repeat(mels_i, 3).reshape((*mels_i.shape, 3))
    # The following should generate the same ndarray, and maybe faster or at least clearer
    #X_train[i] = np.repeat(mels_i[..., np.newaxis], 3, axis=-1)
X_train /= 255.0
X_train.max(), X_train.min()

def birds_to_ndarry(series):
    I = np.eye(len(D_label_index))
    ndarray = np.zeros((len(series), len(D_label_index)))
    for i, string in enumerate(series.values):
        if string == "nocall":
            continue
        else:
            L_indices = [D_label_index[label] for label in string.split(" ")]
            row_i = np.sum(I[L_indices], axis=0)
            ndarray[i] = row_i
    return ndarray

A = birds_to_ndarry(df_train_soundscape.birds)
A.shape

y_train = A[df_train_soundscape.is_test==False]
y_train.shape

y_test = A[df_train_soundscape.is_test==True]
y_test.shape

## Models
- EfficientNet

from tensorflow.keras.applications import EfficientNetB0

(*random_npy.shape, 3)

help(EfficientNetB0)

input_mels = keras.layers.Input(shape=(*random_npy.shape, 3), name="input_mels")
input_spacetime = keras.layers.Input(shape=(XX_train_normalized.shape[1],),
                                     name="input_spacetime")

output_efficient = EfficientNetB0(include_top=False, weights="imagenet")(input_mels)

[s for s in dir(output_efficient) if not s.startswith("_")]

output_efficient.type_spec

`(4, 6, 1280)` weird shape... How is the structure of EfficientNet?

m = EfficientNetB0(weights='imagenet')
m.summary()

I guess we have this weird shape of `(4, 6, 1280)` because of our own fault: We had `Input` with shape `(128, 201, 3)` while it seems that
EfficientNet expects square images (shape `(224, 224, 3)`).<br>
_However, it seems that `kkiller` used Resnest with this same weird shape without reshaping_.<br>
Let's explore a little bit.

pooled = keras.layers.GlobalAveragePooling2D()(output_efficient)
pooled.shape

concatenated = keras.layers.Concatenate()([pooled, input_spacetime])

#dropped = keras.layers.Dropout(.2)(pooled)
dropped = keras.layers.Dropout(.2)(concatenated)
dropped.shape

output_CNN = keras.layers.Dense(len(L_birds), activation="sigmoid")(dropped)
output_CNN.shape

dir(keras.metrics)

model_image_only = keras.Model(
    #inputs=[input_mels],
    inputs=[input_mels, input_spacetime],
    outputs=[output_CNN],
)
#model_image_only.compile(loss="binary_crossentropy", optimizer="adam", metrics=["acc"])
model_image_only.compile(
    loss="binary_crossentropy",
    optimizer="adam",
    metrics=[keras.metrics.Precision(), keras.metrics.Recall()],
)

help(model_image_only.fit)

checkpoint_cb = keras.callbacks.ModelCheckpoint("model1.h5",
                                                save_best_only=True)
early_stopping_cb = keras.callbacks.EarlyStopping(patience=10,
                                                  restore_best_weights=True)

class PrintF1Score(keras.callbacks.Callback):
    def on_epoch_end(self, epoch, logs):
        #print(f"logs.keys() = {logs.keys()}")  # This can check what keys logs has.
        f1_score = 2 * logs["precision"] * logs["recall"] / (logs["precision"] + logs["recall"] + EPSILON)
        val_f1_score = 2 * logs["val_precision"] * logs["val_recall"] / (logs["val_precision"] + logs["val_recall"] + EPSILON)
        print(f"f1_score: {f1_score}")
        print(f"val_f1_score: {val_f1_score}")

history = model_image_only.fit(
    (X_train, XX_train_normalized),
    y_train,
    batch_size=32,
    epochs=100,
    callbacks=[checkpoint_cb, early_stopping_cb, PrintF1Score()],
    validation_split=0.2,
)

## Test Set

n_test_instances = df_train_soundscape.loc[is_test].shape[0]
X_test = np.empty((n_test_instances, *random_npy.shape, 3), dtype=np.float32)
print(f"X_test.shape = {X_test.shape}")

# fill in X_test
for i, row in enumerate((df_train_soundscape.loc[is_test]).itertuples()):
    mels_i = np.load(row.npy_path).astype(np.float32, copy=False)
    #X_test[i] = np.repeat(mels_i, 3).reshape((*mels_i.shape, 3))
    # The following should generate the same ndarray, and maybe faster or at least clearer
    X_test[i] = np.repeat(mels_i[..., np.newaxis], 3, axis=-1)
X_test /= 255.0
X_test.max(), X_test.min()
