==> kaggle/birdclef_2021/my_code/local/07-debug_dataset_generator.ipynb <==
import tensorflow as tf
import tensorflow.keras as keras
import numpy as np

dir(tf.random)

def trainset_generator():
    for _ in range(100):
        input1 = tf.random.uniform((7,))
        input2 = tf.random.uniform((5,))
        y = tf.zeros((1,))
        yield [input1, input2], y

i1 = keras.layers.Input(shape=(7,))
i2 = keras.layers.Input(shape=(5,))
d1 = keras.layers.Dense(200)(i1)
o = keras.layers.Concatenate()([d1, i2])
m = keras.Model(
    inputs=[i1, i2],
    outputs=[o],
)
m.summary()

m.compile(
    loss="binary_crossentropy",
    optimizer="adam",
)

m.fit(trainset_generator(), batch_size=32)

ValueError: Input 0 of layer dense_1 is incompatible with the layer: expected axis -1 of input shape to have value 7 but received input with shape (None, 1)

from tensorflow.keras.utils import Sequence
class DatasetGenerator(Sequence):
    def __init__(self, batch_size=32):
        self.batch_size = batch_size
        self.n_instances = int(1e6)
    def __len__(self):
        return self.n_instances // self.batch_size
    def __getitem__(self, idx):
        input1 = tf.random.uniform((32, 7))
        input2 = tf.random.uniform((32, 5))
        batch_x = (input1, input2)
        batch_y = tf.zeros((32,))
        #y = tf.zeros((32,1))
        return batch_x, batch_y

m.fit(DatasetGenerator(), batch_size=32)
